requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
openai==1.54.5
python-multipart==0.0.18
pydantic==2.10.3
pillow==11.0.0
pytest==8.3.4
httpx==0.27.2
requests==2.32.3
